add missing  SPARK-38124 + remove rangeBase class
